fixed the delete bug

+++ controllers/V1controllers/V2ProjectController.js
@@ -26,14 +26,23 @@ router.get("/", (req, res) => {
 
 //*--* Update functionality for after V3. *--*//
 
-//Delete one project by ID
+//Delete one sting by ID
 router.delete("/:id", authenticate, (req, res) => {
-  Project.findByIdAndDelete(req.params.id).then((project) => {
-    res.json({
-      status: 200,
-      msg: "Item deleted.",
-      project: project,
-    });
+  Project.findById(req.params.id).then((project) => {
+    const data = project.directorID;
+    const userID = req.user.id;
+    if (data == userID) {
+      Project.findByIdAndDelete(req.params.id).then((project) => {
+        res.json({
+          status: 200,
+          msg: "Item deleted.",
+          project: project,
+          id: data,
+        });
+      });
+    } else {
+      console.log("Cannot delete.");
+    }
   });
 });
